step1: refactor testOrg creation
https://coveord.atlassian.net/browse/CDX-860
